[WIP] Completes bindings for libmemcached.vapi.
Posix is required as the 'memcached_stat_st' struct has a 'pid_t'
attribute.

Finishes the bindings for 'result.h' and cleans the wrapper code.

Defines ulonglong IntegerType to produce 'unsigned long long' for the
stats struct.

dump, fetch_execute, mget_execute and mget_execute_by_key take a
sequence of callback and expect a void* context to be passed. As this is
not supported by Vala (multiple delegates sharing the same context), it
is bound by:

 - using a ref on a single function (generates a pointer on a function)
 - defaulting number_of_callbacks to '1'